apps/telegram_assistant/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
aiogram==3.4.1
httpx==0.26.0
pydantic==2.6.3
pydantic-settings==2.2.1
supabase==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.1
